Keep utils in a separate util file
